Restructured test files
